added some fx:ids, renamed the main file, and added a public link to sql db
