+++ src/main/java/module-info.java
@@ -1,8 +1,9 @@
 module com.example.recipefinder {
     requires javafx.controls;
     requires javafx.fxml;
-            
-                            
+    requires java.sql;
+
+
     opens com.example.recipefinder to javafx.fxml;
     exports com.example.recipefinder;
 }
